pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "densa-deck"
version = "0.1.0"
description = "MTG deck analysis, goldfish testing, matchup simulation, and player insight engine"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "httpx>=0.27",
    "rich>=13.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]
desktop = [
    "pyinstaller>=6.0",
    # GUI shell for the desktop app (Stage A). pywebview wraps a native
    # OS webview + exposes a Python API to the HTML/JS frontend. Smaller
    # than Electron (~20 MB overhead vs ~150 MB) and doesn't need a
    # separate browser runtime. Optional — CLI still works without it.
    "pywebview>=5.0",
]
analyst = [
    # Local GGUF inference for the LLM analyst layer. Optional — the analyst
    # falls back to MockBackend if this isn't installed.
    "llama-cpp-python>=0.3",
]

[project.scripts]
densa-deck = "densa_deck.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
# Bundle the desktop app's HTML/CSS/JS so it ships inside the wheel and
# the PyInstaller frozen binary finds them at STATIC_DIR.
# Note: static/*.js covers both app.js and tour.js.
"densa_deck.app" = ["static/*.html", "static/*.css", "static/*.js"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
